+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Technical Changes
 
+## 8.0.3 (2023-04-xx)
+
+### linux.py
+- Added  to `XErrorEvent` class (`Event` will be removed in v9.0.0)
+
 ## 8.0.0 (2023-04-09)
 
 ### base.py